Fix #23 Organize ChatView files
